+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 2023-12-16
+## [Unreleased] 2023-12-17
 ### Changed
 - Migration of simpleapiRoute to TS
+- Migration of deleteTaskRoute to TS
+- Migration of getSecretsRoute to TS
+- Migration of sessionRoute to TS
+- Migration of switchTaskEnablerRoute to TS
+
+### Fixed
+- Fixed the association between zoneName and taskId to ensure the correct task is deleted.
 
 ## [v15.4.0] 2023-12-10
 ### Changed